requirements.txt
streamlit
google-generativeai
python-pptx
python-docx
python-dotenv
pillow
pandas
bcrypt